+++ src/agent_skills/__init__.py
@@ -7,10 +7,11 @@ from .models import (
     ValidationResult,
 )
 from .parser import SkillParseError, parse_skill_file
+from .reporting import inspect_to_payload, summaries_to_payload, validation_to_payload
 from .service import discover_skills, inspect_skill, validate_skill, validate_skills
 from .validator import validate_skill_document
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"
 
 __all__ = [
     "__version__",
@@ -28,4 +29,7 @@ __all__ = [
     "validate_skill",
     "validate_skills",
     "discover_skills",
+    "summaries_to_payload",
+    "inspect_to_payload",
+    "validation_to_payload",
 ]